Moved express app to functions index
