+++ src/components/dashboard/Billing.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import { Link } from '@reach/router';
 import moment from 'moment-timezone';
 import styled from 'styled-components';
+import axios from 'axios';
 import Loading from '../Loading';
 import { db } from '../../firebase';
 
@@ -12,44 +13,13 @@ const Billing = ({ user }) => {
   useEffect(() => {
     const fetchData = async () => {
       try {
-        const response = await fetch(
-          `https://api.stripe.com/v1/customers/${user.stripe_id}`,
-          {
-            method: 'GET',
-            headers: {
-              'Content-Type': 'application/x-www-form-urlencoded',
-              Authorization: `Bearer ${process.env.REACT_APP_STRIPESECRET}`,
-            },
-          },
+        const response = await axios.get(
+          `https://us-central1-recaller-14a1f.cloudfunctions.net/stripe/billing/${
+            user.stripe_id
+          }`,
         );
-        const data = await response.json();
-        const subs = data.subscriptions.data.map(async sub => {
-          const contact = await db
-            .doc(`/contacts/${sub.metadata.contact_id}`)
-            .get();
-          const user2 = await contact.data().user2.get();
-          const invoicesFetch = await fetch(
-            `https://api.stripe.com/v1/invoices?subscription=${sub.id}`,
-            {
-              method: 'GET',
-              headers: {
-                'Content-Type': 'application/x-www-form-urlencoded',
-                Authorization: `Bearer ${process.env.REACT_APP_STRIPESECRET}`,
-              },
-            },
-          );
-          const invoices = await invoicesFetch.json();
-          return {
-            contact_id: contact.id,
-            user2: user2.data(),
-            contact: { ...contact.data(), id: contact.id },
-            next_charge_date: sub.current_period_end,
-            amount: sub.plan.amount,
-            invoices: invoices.data,
-          };
-        });
-        const results = await Promise.all(subs);
-        setSubs(s => results);
+        console.log(response.data);
+        // setSubs(s => results);
         setIsLoading(false);
       } catch (err) {
         console.log(err);
